demux_lavf: apply hacks even if format is forced
Some of the hacks were not applied if the file format was forced. Commit
5fb18a75 moved them to a table, which is checked with normal probing
only.

Fixes #1612 (DVD forces mpeg, which in turn has to export native stream
IDs specifically).

Do some code restructuring on the way. For example, the probescore can
simply be set to the correct initial value, instead of checking whether
it was set at all.